# src/main/resources/static/api.yaml
openapi: 3.0.1
info:
  title: SerieApi
  description: >-
    API REST para gerenciamento de séries, plataformas de streaming e usuários.
    Permite cadastrar, atualizar, filtrar e associar séries a plataformas de streaming.
  version: v1.0.0
  contact:
    name: Suporte SerieApi
    email: suporte@serieapi.com.br
security:
  - bearerAuth: [ ]
servers:
  - url: http://localhost:8080
    description: Ambiente de desenvolvimento local
...
paths:
...
  /series/{serieId}/ativar:
    put:
      tags:
        - Series
      summary: Ativa uma série
      description: Altera o status da série para ativo, indicando que ainda recebe novos episódios.
      operationId: ativarSerie
      parameters:
        - $ref: '#/components/parameters/serieId'
      responses:
        '204':
          description: Série ativada com sucesso
        '404':
          description: Série não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'
  /series/{serieId}/desativar:
    delete:
      tags:
        - Series
      summary: Desativa uma série
      description: Altera o status da série para inativo, indicando que não recebe mais novos episódios.
      operationId: desativarSerie
      parameters:
        - $ref: '#/components/parameters/serieId'
      responses:
        '204':
          description: Série desativada com sucesso
        '404':
          description: Série não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'

  #  /series/filtros:
  #    get:
  #      tags:
  #        - Series
  #      summary: Filtra séries dinamicamente
  #      description: Busca séries aplicando filtros opcionais combinados. Todos os parâmetros são opcionais.
  #      operationId: listarSeriesComFiltros
  #      parameters:
  #        - name: titulo
  #          in: query
  #          description: Filtra séries pelo título (busca parcial)
  #          required: false
  #          schema:
  #            type: string
  #          example: Breaking
  #        - name: notaMax
  #          in: query
  #          description: Filtra séries com avaliação até o valor informado
  #          required: false
  #          schema:
  #            type: number
  #            format: double
  #          example: 9.5
  #        - name: limite
  #          in: query
  #          description: Limita a quantidade de resultados retornados
  #          required: false
  #          schema:
  #            type: integer
  #            format: int32
  #          example: 10
  #      responses:
  #        '200':
  #          description: Filtro processado com sucesso
  #          content:
  #            application/json:
  #              schema:
  #                type: array
  #                items:
  #                  $ref: '#/components/schemas/SerieResponseDto'
  #        '500':
  #          description: Erro interno no servidor
  #          content:
  #            application/json:
  #              schema:
  #                $ref: '#/components/schemas/ErroInternoResponse'

  # ==========================================
  # SÉRIES x STREAMINGS
  # ==========================================
...
  /streamings:
    get:
      tags:
        - Streamings
      summary: Lista todos os streamings
      description: Retorna todas as assinaturas de streaming cadastradas.
      operationId: listarStreamings
      responses:
        '200':
          description: Streamings retornados com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StreamingResponseDto'

    post:
      tags:
        - Streamings
      summary: Cria uma nova assinatura de streaming
      description: Cria o vínculo entre um usuário e uma plataforma de streaming.
      operationId: salvarStreaming
      requestBody:
        description: Dados da assinatura de streaming
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamingRequestDto'
            examples:
              ExemploRequest:
                summary: Exemplo de criação de streaming
                value:
                  usuarioAtivo: true
                  estaNaPlataforma: true
                  dataNaPlataforma: 2024-01-15
                  usuarioId:
                    id: 1
                  plataformaId:
                    id: 1
      responses:
        '201':
          description: Streaming criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamingResponseDto'
        '400':
          description: Dados de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroDadosInvalidosResponse'
  /streamings/{streamingId}:
    get:
      tags:
        - Streamings
      summary: Busca um streaming pelo ID
      description: Retorna os dados de uma assinatura de streaming específica pelo seu ID.
      operationId: buscarStreamingPorId
      parameters:
        - $ref: '#/components/parameters/streamingId'
      responses:
        '200':
          description: Streaming encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamingResponseDto'
        '404':
          description: Streaming não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'

    put:
      tags:
        - Streamings
      summary: Atualiza um streaming pelo ID
      description: Atualiza os dados de uma assinatura de streaming existente.
      operationId: atualizarStreaming
      parameters:
        - $ref: '#/components/parameters/streamingId'
      requestBody:
        description: Novos dados do streaming
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamingRequestDto'
      responses:
        '200':
          description: Streaming atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamingResponseDto'
        '400':
          description: Dados de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroDadosInvalidosResponse'
        '404':
          description: Streaming não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'

    delete:
      tags:
        - Streamings
      summary: Remove um streaming pelo ID
      description: Remove permanentemente uma assinatura de streaming do banco de dados.
      operationId: apagarStreaming
      parameters:
        - $ref: '#/components/parameters/streamingId'
      responses:
        '204':
          description: Streaming removido com sucesso
        '404':
          description: Streaming não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'

# ==========================================
# COMPONENTS
# ==========================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    serieId:
      name: serieId
      in: path
      description: ID numérico da série
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
    streamingId:
      name: streamingId
      in: path
      description: ID numérico do streaming
      required: true
      schema:
        type: integer
        format: int64
      example: 1
  schemas:
...
    StreamingRequestDto:
      type: object
      description: Dados para criação ou atualização de uma assinatura de streaming
      required:
        - usuarioId
        - plataformaId
      properties:
        usuarioAtivo:
          type: boolean
          description: Indica se o usuário está ativo na plataforma
          example: true
        estaNaPlataforma:
          type: boolean
          description: Indica se o usuário ainda possui assinatura ativa
          example: true
        dataNaPlataforma:
          type: string
          format: date
          description: Data de início da assinatura no formato yyyy-MM-dd
          example: 2024-01-15
        usuarioId:
          $ref: '#/components/schemas/UsuarioIdRequestDto'
        plataformaId:
          $ref: '#/components/schemas/PlataformaIdRequestDto'
...
    UsuarioIdRequestDto:
      type: object
      description: Referência ao ID de um usuário já cadastrado
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: ID numérico do usuário já cadastrado no banco de dados
          example: 1
    PlataformaIdRequestDto:
      type: object
      description: Referência ao ID de uma plataforma já cadastrada
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: ID numérico da plataforma já cadastrada no banco de dados
          example: 1

    # ---- RESPONSES ----
    SerieResponseDto:
      type: object
      description: Dados completos de uma série retornados pela API
      required:
        - atores
        - avaliacao
        - dataLancamento
        - poster
        - sinopse
        - titulo
        - totalTemporada
        - categoriaId
      properties:
        id:
          type: integer
          format: int64
          description: ID único da série
          example: 1
        titulo:
          type: string
          description: Título oficial da série
          example: Breaking Bad
        totalTemporada:
          type: integer
          format: int32
          description: Quantidade total de temporadas
          example: 5
        avaliacao:
          type: number
          format: double
          description: Nota de avaliação da série
          example: 9.5
        atores:
          type: string
          description: Principais atores do elenco
          example: Bryan Cranston, Aaron Paul
        poster:
          type: string
          description: URL do poster da série
          example: https://link-da-imagem.com/poster.jpg
        sinopse:
          type: string
          description: Resumo do enredo
          example: Um professor de química se volta para o crime...
        dataLancamento:
          type: string
          format: date
          description: Data do lançamento original
          x-field-extra-annotation: "@com.fasterxml.jackson.annotation.JsonFormat(pattern = \"yyyy-MM-dd\")"
        ativo:
          type: boolean
          description: Indica se a série ainda recebe novos episódios
          example: true
        categoriaId:
          $ref: '#/components/schemas/CategoriaResponseDto'
    CategoriaResponseDto:
      type: object
      description: Dados de uma categoria retornados pela API
      properties:
        id:
          type: integer
          format: int64
          example: 1
        categoria:
          type: string
          example: Drama
    PlataformaResponseDto:
      type: object
      description: Dados de uma plataforma retornados pela API
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: Netflix
        dominio:
          type: string
          example: netflix.com
        totalCatalogo:
          type: integer
          format: int64
          example: 5000
        mensalidade:
          type: number
          format: double
          example: 39.90
    UsuarioResponseDto:
      type: object
      description: Dados de um usuário retornados pela API
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: João Silva
        email:
          type: string
          example: joao@marinha.mil.br
...
    StreamingResponseDto:
      type: object
      description: Dados de uma assinatura de streaming retornados pela API
      properties:
        id:
          type: integer
          format: int64
          example: 1
        usuario:
          $ref: '#/components/schemas/UsuarioResponseDto'
        plataforma:
          $ref: '#/components/schemas/PlataformaResponseDto'
        usuarioAtivo:
          type: boolean
          example: true
        estaNaPlataforma:
          type: boolean
          example: true
        dataNaPlataforma:
          type: string
          format: date
          example: 2024-01-15
...
    ErroDadosInvalidosResponse:
      type: object
      description: Resposta de erro para dados inválidos (400)
      properties:
        type:
          type: string
          example: https://serieapi.com.br/dados-invalidos
        title:
          type: string
          example: Dados inválidos
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Um ou mais campos estão incorretos. Faça o preenchimento correto.
        instance:
          type: string
          example: /series
        fields:
          type: array
          items:
            $ref: '#/components/schemas/CampoInvalidoResponse'
    CampoInvalidoResponse:
      type: object
      description: Detalhe de um campo inválido
      properties:
        fieldName:
          type: string
          example: titulo
        detail:
          type: string
          example: Título da série é obrigatório.
    ErroNaoEncontradoResponse:
      type: object
      description: Resposta de erro para recurso não encontrado (404)
      properties:
        type:
          type: string
          example: https://serieapi.com.br/recurso-nao-encontrado
        title:
          type: string
          example: Recurso não encontrado
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: Não existe uma série cadastrada com o código 99.
        instance:
          type: string
          example: /series/99
...
    ErroInternoResponse:
      type: object
      description: Resposta de erro para falha interna no servidor (500)
      properties:
        type:
          type: string
          example: https://serieapi.com.br/erro-interno
        title:
          type: string
          example: Erro interno no servidor
        status:
          type: integer
          example: 500
        detail:
          type: string
          example: Ocorreu um erro inesperado. Tente novamente mais tarde.
        instance:
          type: string
          example: /series